market_sentiment_batch/ecs_tasks/news_sentiment_task/requirements.txt
sqlalchemy==1.4.32
mysql-connector-python==8.1.0
python_dotenv==1.0.0
requests==2.27.1